Add a link on empty domains tables
