+++ src/registrar/assets/sass/_theme/_links.scss
@@ -1,0 +1,19 @@
+@use "uswds-core" as *;
+
+p, .dotgov-table {
+    a {
+        display: flex;
+        align-items: flex-start;
+        color: color('primary');
+    
+        &:visited {
+            color: color('primary');
+        }
+    
+        .usa-icon {
+            // align icon with x height
+            margin-top: units(0.5);
+            margin-right: units(0.5);
+        }
+    }
+}
