+++ src/registrar/assets/sass/_theme/styles.scss
@@ -10,6 +10,7 @@
 --- Custom Styles ---------------------------------*/
 @forward "base";
 @forward "typography";
+@forward "links";
 @forward "lists";
 @forward "buttons";
 @forward "forms";
